refactor: 기물(폰) 패키지 분리
- Pawn: 기물들 관련 클래스를 구별할 수 있는 `pieces` 디렉토리로 이동
- Color: 기물과 체스판에 공통으로 사용될 수 있는 상수는 `common` 디렉토리로 이동
