+++ src/test/java/chess/pieces/PawnTest.java
@@ -1,7 +1,9 @@
-package chess.domain;
+package chess.pieces;
 
 import static org.assertj.core.api.Assertions.*;
 
+import chess.common.Color;
+import chess.pieces.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
